Education
Education page overhaul. The education throughout school is now in a grid, which has been made to look like multiple circles. Decided it was better for organization to have a style sheet for each page, rather than one really unorganized one. Added an image to the freshman year section. The navs at the top now have a fade in like effect. Small changes to order and alignment for some text.