Small changes to mobile
Slight improvement to the mobile version of the skills page, and started on the mobile version of the education page. Font size gets smaller for the skills page when below a thresh hold. Also noticed that one of the commit pushes was written for an exercise that was done, it doesn't seem to have made any impact and was just the description and message for one of the exercise pushes, on this one