Navigation for smaller sizes
Worked on the navigation for smaller sizes of screen. Drop-down now works, and all pages now scale better for mobile. Navigation now works slightly wrong, when changing colors it applies to the image that indicates a drop down, but not the links themselves. Had to rename the image for purposes of testing, got it to work anyway(in wrong area).